move useDetectNativeToken to @cow
