+++ src/cow-react/modules/swap/hooks/useIsEthFlow.ts
@@ -1,4 +1,4 @@
-import { useDetectNativeToken } from 'state/swap/hooks'
+import { useDetectNativeToken } from '@cow/modules/swap/hooks/useDetectNativeToken'
 
 export function useIsEthFlow(): boolean {
   const { isNativeIn, isWrapOrUnwrap } = useDetectNativeToken()
