+++ src/custom/state/swap/hooks.tsx
@@ -27,13 +27,11 @@ import { BAD_RECIPIENT_ADDRESSES } from '@src/state/swap/hooks'
 import { useGetQuoteAndStatus, useQuote } from '../price/hooks'
 import { registerOnWindow } from 'utils/misc'
 import { useTradeExactInWithFee, useTradeExactOutWithFee, stringToCurrency } from './extension'
-import { DEFAULT_NETWORK_FOR_LISTS } from 'constants/lists'
-import { FEE_SIZE_THRESHOLD, INITIAL_ALLOWED_SLIPPAGE_PERCENT, WETH_LOGO_URI } from 'constants/index'
-import WXDAI_LOGO_URI from 'assets/cow-swap/wxdai.png'
+import { FEE_SIZE_THRESHOLD, INITIAL_ALLOWED_SLIPPAGE_PERCENT } from 'constants/index'
 import TradeGp from './TradeGp'
 
 import { SupportedChainId, SupportedChainId as ChainId } from 'constants/chains'
-import { WRAPPED_NATIVE_CURRENCY as WETH, GpEther as ETHER } from 'constants/tokens'
+import { WRAPPED_NATIVE_CURRENCY as WETH } from 'constants/tokens'
 
 import { isWrappingTrade } from './utils'
 
@@ -472,43 +470,6 @@ export interface NativeCurrenciesInfo {
   isWrappedOut: boolean
   wrappedToken: Token & { logoURI: string }
   native: NativeCurrency
-}
-
-export function useDetectNativeToken() {
-  const { chainId } = useWeb3React()
-  const {
-    [Field.INPUT]: { currencyId: inputCurrencyId },
-    [Field.OUTPUT]: { currencyId: outputCurrencyId },
-  } = useSwapState()
-
-  const input = useCurrency(inputCurrencyId)
-  const output = useCurrency(outputCurrencyId)
-
-  return useMemo(() => {
-    const activeChainId = supportedChainId(chainId)
-    const wrappedToken: Token & { logoURI: string } = Object.assign(
-      WETH[activeChainId || DEFAULT_NETWORK_FOR_LISTS].wrapped,
-      {
-        logoURI: activeChainId === ChainId.GNOSIS_CHAIN ? WXDAI_LOGO_URI : WETH_LOGO_URI,
-      }
-    )
-
-    // TODO: check the new native currency function
-    const native = ETHER.onChain(activeChainId || DEFAULT_NETWORK_FOR_LISTS)
-
-    const [isNativeIn, isNativeOut] = [!!input?.isNative, !!output?.isNative]
-    const [isWrappedIn, isWrappedOut] = [!!input?.equals(wrappedToken), !!output?.equals(wrappedToken)]
-
-    return {
-      isNativeIn,
-      isNativeOut,
-      isWrappedIn,
-      isWrappedOut,
-      wrappedToken,
-      native,
-      isWrapOrUnwrap: (isNativeIn && isWrappedOut) || (isNativeOut && isWrappedIn),
-    }
-  }, [input, output, chainId])
 }
 
 export function useIsFeeGreaterThanInput({ address, chainId }: { address?: string | null; chainId?: ChainId }): {
